requirements.txt
hyperopt==0.2.7
matplotlib==3.9.0
mlflow==2.13.0
kafka-python==2.0.2
pandas==2.2.2
psycopg2==2.9.9
pyarrow==16.1.0
pyspark==3.5.1
python-dotenv==1.0.1
requests==2.31.0
seaborn==0.13.2
scikit-learn==1.4.2
xgboost==2.0.3